Déplacement dans un nouveau fichier de l'analyes XML
